back to long, but use %lu in printf.  silly me.

+++ p2psim/protocols/kademlia.h
@@ -345,28 +345,28 @@ private:
   static unsigned _nkademlias;
 
   // global statistics
-  static long long unsigned _rpc_bytes;
-  static long long unsigned _good_rpcs;
-  static long long unsigned _bad_rpcs;
-  static long long unsigned _ok_by_reaper;
-  static long long unsigned _timeouts_by_reaper;
+  static long unsigned _rpc_bytes;
+  static long unsigned _good_rpcs;
+  static long unsigned _bad_rpcs;
+  static long unsigned _ok_by_reaper;
+  static long unsigned _timeouts_by_reaper;
 
-  static long long unsigned _good_lookups;
-  static long long unsigned _lookup_dead_node;
-  static long long unsigned _ok_failures;
-  static long long unsigned _bad_failures;
+  static long unsigned _good_lookups;
+  static long unsigned _lookup_dead_node;
+  static long unsigned _ok_failures;
+  static long unsigned _bad_failures;
 
   static Time _good_total_latency;
   static Time _good_lookup_latency;
   static Time _good_ping_latency;
-  static long long unsigned _good_timeouts;
+  static long unsigned _good_timeouts;
 
-  static long long unsigned _good_hops;
+  static long unsigned _good_hops;
   static Time _good_hop_latency;
 
   static Time _bad_lookup_latency;
-  static long long unsigned _bad_timeouts;
-  static long long unsigned _bad_hops;
+  static long unsigned _bad_timeouts;
+  static long unsigned _bad_hops;
   static Time _bad_hop_latency;
